requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
vaderSentiment==3.3.2
groq==0.9.0
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.32.2
fake-useragent==1.5.1
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.5.0
wordcloud==1.9.3
Pillow==10.3.0
lxml==5.2.2
httpx<0.28